% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depreciation}
\alias{depreciation}
\title{Vehicle depreciation quantiles by age, powertrain, and vehicle type}
\format{
A tibble with 6,486 rows and 5 variables:\tabular{ll}{
   Variable \tab Description \cr
   \code{age_bin} \tab Midpoint of 3-month age bin in months (e.g., 13.5, 16.5, ..., up to 180) \cr
   \code{powertrain} \tab Powertrain category: "All", "Gasoline", "Battery Electric (BEV)", "BEV (Tesla)", "BEV (Non-Tesla)", "Hybrid Electric (HEV)", "Plug-In Hybrid Electric (PHEV)", "Diesel", "Flex Fuel (E85)" \cr
   \code{vehicle_type} \tab Vehicle type: "All", "Car", "CUV", "SUV", "Pickup", "Minivan" \cr
   \code{quantile} \tab Quantile level (25, 50, or 75) \cr
   \code{rr} \tab Retention rate (price / MSRP) at the given quantile \cr
}
}
\source{
Computed from used vehicle listings data from
\href{https://www.marketcheck.com/}{Marketcheck}.
}
\usage{
data(depreciation)
}
\description{
Retention rate (price / MSRP) quantiles for used vehicles, computed at
3-month age bins from vehicle listings data. Only used vehicles with valid
price and MSRP values are included, with ages ranging from 12 to 180 months
(15 years). Quantiles are provided at the 25th, 50th, and 75th percentiles
across various groupings of powertrain and vehicle type.
}
\examples{
data(depreciation)

head(depreciation)
}
\keyword{datasets}
