% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{percent_dealers}
\alias{percent_dealers}
\title{Percentage of dealers with at least one listing by variable pairs}
\format{
A data frame with 7 variables:\tabular{ll}{
   Variable \tab Description \cr
   \code{listing_year} \tab Year of the vehicle listing (2018--2024) \cr
   \code{inventory_type} \tab Inventory type: "New" or "Used" \cr
   \code{group_var} \tab Name of the grouping variable: "powertrain", "vehicle_type", or "price_bin" \cr
   \code{group_level} \tab Level of the grouping variable (e.g., "Gasoline", "Car", "$30k-$40k") \cr
   \code{category_var} \tab Name of the category variable: "powertrain", "vehicle_type", or "price_bin" \cr
   \code{category_level} \tab Level of the category variable \cr
   \code{p} \tab Proportion of dealers with at least one listing in this group-category combination \cr
}
}
\source{
Computed from vehicle listings data from
\href{https://www.marketcheck.com/}{Marketcheck}.
}
\usage{
data(percent_dealers)
}
\description{
Percentage of dealers that have at least one listing for each combination of
two grouping variables, computed across all listing years and inventory types.
Three variable-pair combinations are included: powertrain by vehicle type,
powertrain by price bin, and vehicle type by price bin.
}
\examples{
data(percent_dealers)

head(percent_dealers)
}
\keyword{datasets}
