% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_trends_new}
\alias{price_trends_new}
\title{Price trends for new vehicle listings by state, year, and vehicle type}
\format{
A tibble with 7 variables:\tabular{ll}{
   Variable \tab Description \cr
   \code{listing_year} \tab Calendar year when listing was scraped (2018--2025) \cr
   \code{state} \tab US state abbreviation (e.g., "CA", "TX", "NY") \cr
   \code{year} \tab Model year of the vehicle \cr
   \code{vehicle_type} \tab Vehicle type: "car", "cuv", "suv", "pickup", "minivan" \cr
   \code{avg_price} \tab Average listing price (USD) \cr
   \code{n_listings} \tab Number of listings \cr
   \code{n_make_model_year} \tab Number of unique make-model-year combinations in this state \cr
}
}
\source{
Computed from new vehicle listings data from
\href{https://www.marketcheck.com/}{Marketcheck}.
}
\usage{
data(price_trends_new)
}
\description{
Average prices and listing counts for new vehicle inventory by listing year,
state, model year, and vehicle type. Includes the number of unique
make-model-year combinations available in each state.
}
\examples{
data(price_trends_new)

head(price_trends_new)
}
\keyword{datasets}
