% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_trends_used}
\alias{price_trends_used}
\title{Price trends for used vehicle listings by state, year, and vehicle type}
\format{
A tibble with 8 variables:\tabular{ll}{
   Variable \tab Description \cr
   \code{listing_year} \tab Calendar year when listing was scraped (2018--2025) \cr
   \code{state} \tab US state abbreviation (e.g., "CA", "TX", "NY") \cr
   \code{year} \tab Model year of the vehicle \cr
   \code{vehicle_type} \tab Vehicle type: "car", "cuv", "suv", "pickup", "minivan" \cr
   \code{avg_price} \tab Average listing price (USD) \cr
   \code{n_listings} \tab Number of listings \cr
   \code{dep_annual_rate} \tab Annual depreciation rate (proportion, 0--1) for this state and listing year \cr
   \code{n_make_model_year} \tab Number of unique make-model-year combinations in this state \cr
}
}
\source{
Computed from used vehicle listings data from
\href{https://www.marketcheck.com/}{Marketcheck}.
}
\usage{
data(price_trends_used)
}
\description{
Average prices, listing counts, and depreciation rates for used vehicle
inventory by listing year, state, model year, and vehicle type. Includes
annual depreciation rates calculated using exponential decay modeling on
retention rates (price/MSRP) and the number of unique make-model-year
combinations available in each state.
}
\examples{
data(price_trends_used)

head(price_trends_used)
}
\keyword{datasets}
