% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vmt_annual_mm}
\alias{vmt_annual_mm}
\title{Estimated annual VMT by make and model}
\format{
A data frame with 551 rows and 5 variables:\tabular{ll}{
   Variable \tab Description \cr
   \code{make} \tab Vehicle manufacturer (e.g., "Toyota", "BMW", "Mercedes-Benz") \cr
   \code{model} \tab Vehicle model name (e.g., "Camry", "RAV4", "Model 3") \cr
   \code{vehicle_type} \tab Vehicle type: "Car", "CUV", "SUV", "Pickup", "Minivan" \cr
   \code{powertrain} \tab Powertrain category: "Gasoline", "Battery Electric (BEV)", "Hybrid Electric (HEV)", "Plug-In Hybrid Electric (PHEV)", "Diesel", "Flex Fuel (E85)" \cr
   \code{vmt_annual} \tab Estimated annual vehicle miles traveled \cr
}
}
\source{
Computed from used vehicle listings data from
\href{https://www.marketcheck.com/}{Marketcheck}.
}
\usage{
data(vmt_annual_mm)
}
\description{
Estimated annual vehicle miles traveled for individual vehicle make/model
combinations. This is an alternate name for the same data as
\link{vmt_annual_model}.
}
\examples{
data(vmt_annual_mm)

head(vmt_annual_mm)
}
\keyword{datasets}
