% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vmt_annual_pt_vt}
\alias{vmt_annual_pt_vt}
\title{Estimated annual VMT by powertrain and vehicle type}
\format{
A data frame with 31 rows and 3 variables:\tabular{ll}{
   Variable \tab Description \cr
   \code{vehicle_type} \tab Vehicle type: "Car", "CUV", "SUV", "Pickup", "Minivan" \cr
   \code{powertrain} \tab Powertrain category: "Gasoline", "Battery Electric (BEV)", "Hybrid Electric (HEV)", "Plug-In Hybrid Electric (PHEV)", "Diesel", "Flex Fuel (E85)" \cr
   \code{vmt_annual} \tab Estimated annual vehicle miles traveled \cr
}
}
\source{
Computed from used vehicle listings data from
\href{https://www.marketcheck.com/}{Marketcheck}.
}
\usage{
data(vmt_annual_pt_vt)
}
\description{
Estimated annual vehicle miles traveled for each combination of powertrain
and vehicle type. This is an alternate name for the same data as
\link{vmt_annual_type}.
}
\examples{
data(vmt_annual_pt_vt)

head(vmt_annual_pt_vt)
}
\keyword{datasets}
