% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vmt_daily}
\alias{vmt_daily}
\title{Daily vehicle miles traveled quantiles}
\format{
A tibble with 5,346 rows and 4 variables:\tabular{ll}{
   Variable \tab Description \cr
   \code{powertrain} \tab Powertrain category: "All", "Gasoline", "Battery Electric (BEV)", "BEV (Tesla)", "BEV (Non-Tesla)", "Hybrid Electric (HEV)", "Plug-In Hybrid Electric (PHEV)", "Diesel", "Flex Fuel (E85)" \cr
   \code{vehicle_type} \tab Vehicle type: "All", "Car", "CUV", "SUV", "Pickup", "Minivan" \cr
   \code{quantile} \tab Quantile level (1 through 99) \cr
   \code{miles} \tab Daily vehicle miles traveled at the given quantile \cr
}
}
\source{
Computed from used vehicle listings data from
\href{https://www.marketcheck.com/}{Marketcheck}.
}
\usage{
data(vmt_daily)
}
\description{
Daily VMT quantiles for used vehicles, computed as total odometer miles
divided by vehicle age in days. Only used vehicles with valid mileage and
age up to 15 years are included. Percentile quantiles from the 1st through
99th are provided across various groupings of powertrain and vehicle type.
}
\examples{
data(vmt_daily)

head(vmt_daily)
}
\keyword{datasets}
